+++ packages/tools/src/types/InterpolationTypes.ts
@@ -2,12 +2,16 @@ import { Types } from '@cornerstonejs/core';
 import { Annotation } from './AnnotationTypes';
 import { InterpolationROIAnnotation } from './ToolSpecificAnnotationTypes';
 
+/**
+ * A base viewport and annotation information used to start interpolating
+ * between slices.
+ */
 export type InterpolationViewportData = {
-  /** The annotation that is being updated with a change in label. */
+  /** The annotation that was just completed. */
   annotation: InterpolationROIAnnotation;
   /** The type of event, whether initializing the label or updating it. */
   interpolationUID: string;
-  /** unique id of the viewport */
+  /** The viewport that this interpolation is occurring within */
   viewport: Types.IViewport;
   sliceData: Types.ImageSliceData;
 };
@@ -15,4 +19,24 @@ export type InterpolationViewportData = {
 export type ImageInterpolationData = {
   sliceIndex: number;
   annotations?: Annotation[];
+};
+
+/**
+ * The selector object for accepting interpolation results.  This object
+ * can be specified to select a sub-set of interpolation results.
+ */
+export type AcceptInterpolationSelector = {
+  /**
+   *  Specify the tool names to apply this to, defaulting to all
+   *  interpolation tools registered
+   */
+  toolNames?: string[];
+  /**
+   * Applies just to the given segmentation
+   */
+  segmentationId?: string;
+  /**
+   * Applies just to the given segment index.
+   */
+  segmentIndex?: number;
 };